# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Text-to-Learn API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string;
  username?: string;
}

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: AuthUser;
}

export interface GenerateCourseBody {
  /**
   * @minLength 3
   * @maxLength 500
   */
  topic: string;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  topic: string;
  tags: string[];
  moduleCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface LessonSummary {
  id: number;
  title: string;
  orderIndex: number;
  isGenerated: boolean;
}

export interface ModuleWithLessons {
  id: number;
  title: string;
  orderIndex: number;
  lessons: LessonSummary[];
}

export interface CourseWithModules {
  id: number;
  title: string;
  description: string;
  topic: string;
  tags: string[];
  modules: ModuleWithLessons[];
  createdAt: string;
  updatedAt: string;
}

export type ContentBlockType =
  (typeof ContentBlockType)[keyof typeof ContentBlockType];

export const ContentBlockType = {
  heading: "heading",
  paragraph: "paragraph",
  code: "code",
  video: "video",
  mcq: "mcq",
} as const;

export interface ContentBlock {
  type: ContentBlockType;
  text?: string;
  language?: string;
  url?: string;
  query?: string;
  question?: string;
  options?: string[];
  answer?: number;
  explanation?: string;
}

export interface LessonWithContent {
  id: number;
  title: string;
  objectives: string[];
  content: ContentBlock[];
  isGenerated: boolean;
  moduleId: number;
  moduleTitle: string;
  courseId: number;
  courseTitle: string;
  orderIndex: number;
}

export interface YoutubeSearchResult {
  videoId: string;
  title: string;
  thumbnail: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export type SearchYoutubeParams = {
  query: string;
};