# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Text-to-Learn API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: courses
    description: Course management
  - name: lessons
    description: Lesson management
  - name: youtube
    description: YouTube video search
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user or unauthenticated state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"

  /courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: Get all courses for authenticated user
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "401":
          description: Unauthorized
    post:
      operationId: generateCourse
      tags: [courses]
      summary: Generate a new course from a topic prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized

  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course with modules and lessons
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseWithModules"
        "404":
          description: Not found
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found

  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a single lesson with full content
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonWithContent"
        "404":
          description: Not found

  /lessons/{lessonId}/generate:
    post:
      operationId: generateLessonContent
      tags: [lessons]
      summary: Generate AI content for a lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson with generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonWithContent"
        "404":
          description: Not found

  /youtube/search:
    get:
      operationId: searchYoutube
      tags: [youtube]
      summary: Search YouTube for educational videos
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: YouTube search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoutubeSearchResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    GetCurrentAuthUserResponse:
      type: object
      properties:
        isAuthenticated:
          type: boolean
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - isAuthenticated

    AuthUser:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: string
        username:
          type: string
      required:
        - id

    GenerateCourseBody:
      type: object
      properties:
        topic:
          type: string
          minLength: 3
          maxLength: 500
      required:
        - topic

    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        topic:
          type: string
        tags:
          type: array
          items:
            type: string
        moduleCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - topic
        - tags
        - moduleCount
        - createdAt
        - updatedAt

    CourseWithModules:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        topic:
          type: string
        tags:
          type: array
          items:
            type: string
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleWithLessons"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - topic
        - tags
        - modules
        - createdAt
        - updatedAt

    ModuleWithLessons:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        orderIndex:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/LessonSummary"
      required:
        - id
        - title
        - orderIndex
        - lessons

    LessonSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        orderIndex:
          type: integer
        isGenerated:
          type: boolean
      required:
        - id
        - title
        - orderIndex
        - isGenerated

    LessonWithContent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        objectives:
          type: array
          items:
            type: string
        content:
          type: array
          items:
            $ref: "#/components/schemas/ContentBlock"
        isGenerated:
          type: boolean
        moduleId:
          type: integer
        moduleTitle:
          type: string
        courseId:
          type: integer
        courseTitle:
          type: string
        orderIndex:
          type: integer
      required:
        - id
        - title
        - objectives
        - content
        - isGenerated
        - moduleId
        - moduleTitle
        - courseId
        - courseTitle
        - orderIndex

    ContentBlock:
      type: object
      properties:
        type:
          type: string
          enum: [heading, paragraph, code, video, mcq]
        text:
          type: string
        language:
          type: string
        url:
          type: string
        query:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
        answer:
          type: integer
        explanation:
          type: string
      required:
        - type

    YoutubeSearchResult:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        thumbnail:
          type: string
      required:
        - videoId
        - title
        - thumbnail

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
          format: uri
        state:
          type: string
        nonce:
          type: string
          nullable: true
      required:
        - code
        - code_verifier
        - redirect_uri
        - state

    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    LogoutMobileSessionResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success