# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Text-to-Learn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string;
  username?: string;
}