# lib/api-zod/src/generated/types/contentBlock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Text-to-Learn API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentBlockType } from "./contentBlockType";

export interface ContentBlock {
  type: ContentBlockType;
  text?: string;
  language?: string;
  url?: string;
  query?: string;
  question?: string;
  options?: string[];
  answer?: number;
  explanation?: string;
}