# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Text-to-Learn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  title: string;
  description: string;
  topic: string;
  tags: string[];
  moduleCount: number;
  createdAt: Date;
  updatedAt: Date;
}