# lib/api-zod/src/generated/types/courseWithModules.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Text-to-Learn API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleWithLessons } from "./moduleWithLessons";

export interface CourseWithModules {
  id: number;
  title: string;
  description: string;
  topic: string;
  tags: string[];
  modules: ModuleWithLessons[];
  createdAt: Date;
  updatedAt: Date;
}