# lib/api-zod/src/generated/types/generateCourseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Text-to-Learn API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateCourseBody {
  /**
   * @minLength 3
   * @maxLength 500
   */
  topic: string;
}