# lib/api-zod/src/generated/types/lessonWithContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Text-to-Learn API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentBlock } from "./contentBlock";

export interface LessonWithContent {
  id: number;
  title: string;
  objectives: string[];
  content: ContentBlock[];
  isGenerated: boolean;
  moduleId: number;
  moduleTitle: string;
  courseId: number;
  courseTitle: string;
  orderIndex: number;
}